+++ src/extension.ts
@@ -10,6 +10,7 @@ import { MarkdownIt } from 'markdown-it';
 import { PuppeteerHelper } from './helpers/PuppeteerHelper';
 import { ExplorerHelper } from './helpers/ExplorerHelper';
 import { ExtensionHelper } from './helpers/ExtensionHelper';
+import { DMBSettings } from './Settings';
 
 interface ContextProperties {
     localStoragePath: string;
@@ -96,7 +97,7 @@ export async function activate(context: ExtensionContext) {
 
     let onEnabledChangeListener = workspace.onDidChangeConfiguration(async cfg => {
         if (cfg.affectsConfiguration('dmbinder.homebrewPreviewEnabled')) {
-            await ExtensionHelper.assertMarkdownPreviewStyles();
+            await ExtensionHelper.toggleHomebreweryStyles(DMBSettings.homebreweryEnabled || false);
             commands.executeCommand('markdown.preview.refresh', undefined);
         }
     });
@@ -117,7 +118,7 @@ export async function activate(context: ExtensionContext) {
     context.subscriptions.push(commands.registerCommand('dmbinder.component.addFile', ExplorerHelper.addNewTreeItem(contextProps.componentsTreeView, CampaignItemType.Component)));
     context.subscriptions.push(commands.registerCommand('dmbinder.generator.addFile', ExplorerHelper.addNewTreeItem(contextProps.generatorsTreeView, CampaignItemType.Generator)));
 
-    await ExtensionHelper.assertMarkdownPreviewStyles();
+    await ExtensionHelper.assertMarkdownStyleDirExists();
 
     return {
         extendMarkdownIt(md: MarkdownIt) {